fix: DeckCardDetailView opacity 타입 모호성 해결
Int - CGFloat 연산 모호성 → 1.0 + Double 캐스팅으로 해결

+++ Momentum Clash/Views/DeckCardDetailView.swift
@@ -15,7 +15,7 @@ struct DeckCardDetailView: View {
         ZStack {
             // 스와이프 시 뒤에 보이는 배경 (검정 → 투명)
             Color.black
-                .opacity(max(0, 1 - dragOffset / 400))
+                .opacity(Double(max(0, 1.0 - dragOffset / 400)))
                 .ignoresSafeArea()
 
             ZStack {
